Refactoring → put all business exception to single package
