+++ src/main/java/hannaeismant/m101j/controllers/SignupController.java
@@ -3,11 +3,11 @@ package hannaeismant.m101j.controllers;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import hannaeismant.m101j.TemplateConfiguration;
-import hannaeismant.m101j.TimeoutException;
-import hannaeismant.m101j.UnknownException;
+import hannaeismant.m101j.exceptions.TimeoutException;
+import hannaeismant.m101j.exceptions.UnknownException;
 import hannaeismant.m101j.session.SessionService;
 import hannaeismant.m101j.session.SessionTokenGenerator;
-import hannaeismant.m101j.user.UserAlreadyExistException;
+import hannaeismant.m101j.exceptions.UserAlreadyExistException;
 import hannaeismant.m101j.user.UserService;
 import spark.Request;
 import spark.Response;
