+++ src/main/java/hannaeismant/m101j/exceptions/TimeoutException.java
@@ -1,4 +1,4 @@
-package hannaeismant.m101j;
+package hannaeismant.m101j.exceptions;
 
 public class TimeoutException extends Exception {
 }
