+++ src/main/java/hannaeismant/m101j/exceptions/UnknownException.java
@@ -1,4 +1,4 @@
-package hannaeismant.m101j;
+package hannaeismant.m101j.exceptions;
 
 public class UnknownException extends Exception {
