+++ src/main/java/hannaeismant/m101j/user/UserService.java
@@ -1,7 +1,8 @@
 package hannaeismant.m101j.user;
 
-import hannaeismant.m101j.TimeoutException;
-import hannaeismant.m101j.UnknownException;
+import hannaeismant.m101j.exceptions.TimeoutException;
+import hannaeismant.m101j.exceptions.UnknownException;
+import hannaeismant.m101j.exceptions.UserAlreadyExistException;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.constraints.NotNull;
